requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.
tornado==4.4.2
oslo.config==6.12.0
oslo.policy==2.4.1
pymongo==3.9.0
PasteDeploy==2.0.1
Werkzeug==0.15.5
keystonemiddleware==7.0.1
requests
